Moving definition of StateSpace into MotionModel.h
