+++ CentralComputing/MotionModel.h
@@ -1,30 +1,34 @@
 #ifndef MOTIONMODEL_H_
 #define MOTIONMODEL_H_
 
-#include "SourceManagerBase.hpp"
-#include "SourceManager.h"
 #include "PRUManager.h"
 #include "ADCManager.h"
-#include "StateSpace.hpp"
 #include "Filter.h"
+#include "Simulator.h"
 
+struct StateSpace {
+  double x[3];  // Position/Velocity/Accel
+  // x[0] = x
+  // x[1] = x' first derivative
+  // x[2] = x'' second derivative
+  
+  double rpm;  // RPM of disks
 
-class MotionModel : public SourceManagerBase<StateSpace, false> {
- private:
-    // Source Manager methods
-    bool initialize_source();
-    void stop_source();
-    std::shared_ptr<StateSpace> refresh();
-    std::shared_ptr<StateSpace> refresh_sim();
-
-    std::string name() {
-      return "mm";
-    }
-
-    // Store our state
-    StateSpace state;
-    int64_t last_time;
+  double fM;  // Force Motor
+  double fD;  // Force Drag
 };
+
+namespace MotionModel {
+  // Source Manager methods
+  bool initialize_source();
+  void stop_source();
+  std::shared_ptr<StateSpace> refresh();
+  std::shared_ptr<StateSpace> refresh_sim();
+
+  // Store our state
+  extern StateSpace state;
+  extern int64_t last_time;
+}  // namepsace MotionModel
 
 #endif  // MOTIONMODEL_H_
